bad refactoring attempt: make Artist a Singleton
Attempt to fix the problem of using a global reference to Artist.first
as a way to make the app single-usered. Do this by giving Artist
Singleton-like behavior.

At first, I didn't realize that I was giving Artist Singleton-like
behavior. When I realized this, I also quickly realized that it was no
better than my previous usage of a global variable. As a result, I am
going to do away with most of the code in this commit. The silver lining
is that I now have a better understanding of why globals are bad and how
they've been causing me all sorts of problems I didn't notice before.