fix Gemfile and features/ and create Admin model
Add "require: false" after cucumber-rails in Gemfile to stop the
"cucumber-rails required outside of env.rb" message.

Move files out of subdirectories in features/ and
features/step_definitions. This was breaking Cucumber's ability to find
step definitions for features when individual features were run. The
app is not large enough to require organizational subdirectories for
features anyway.

Create the basic code for the Admin model. It currently has a username
and a password.