get authentication working again, through Admin!
Successfully move authentication out of Artist and into Admin. This is a
major step for proper separation of concerns between Artist and Admin.

Now, instead of tying the app to a global representation of the first
Artist object in the table, the app implements single-user functionality
by auto-creating an Admin with the login information from
config/admin.yml (if such an Admin does not already exist in the
database) when the user accesses the sign-in page. However, the Admin
constructor just requires a username and password, not a YAML file. This
allows for the creation of mock Admins in tests instead of coupling the
tests to the contents of config/admin.yml. It also will make creating a
multi-user version of the app fairly trivial if that ever becomes
desirable.