set up Admin external flag
Admins now have an :external attribute that specifies whether or not they are
the most recent admin created from config/admin.yml. Only one admin can
have it set to true in the database.

Admin.external_config_admin now searches for an admin with external ==
true in the database unless config/admin.yml exists. If it does, and the
admin account specified in it is valid, it saves that admin to the
database, looks for an admin with :external set to true and toggles that
admin's :external to false if it exists, and sets the new admin's
:external to true.